feat(decorators): property inject input data from Mixer package config
Previously we didn't really use input metadata that much in the controls,
and rather than using the statically-analyzed data we would just get the
control options passed in in the literal decorator. However, this doesn't
work for default values, since default property values are not exposed in
the runtime decorator data.

This does a little bit of refactoring so that input data (and options in
general) are pulled from the package config and overwrite those set at runtime.
I don't entirely eliminate runtime linking in case someone wants to do some
custom hooking system that the static analyzer doesn't see.